Updated LabelSelector to deal with corresponding same as other jsx components
